Save to File

+++ ImageSomCompressor.Core/Som/Neuron/INeuron.cs
@@ -6,9 +6,9 @@ namespace ImageSomCompressor.Core.Som.Neuron
     {
         int X { get; set; }
         int Y { get; set; }
-        int R { get; }
-        int G { get; }
-        int B { get; }
+        byte R { get; }
+        byte G { get; }
+        byte B { get; }
         IVector Weights { get; }
 
         double Distance(INeuron neuron);
